- move keystore files to src folder - move generated d.ts to src folder - simplify closure
