+++ wasm/index.ts
@@ -6,12 +6,10 @@
 
 import * as Loader from "./lib/wallet-core";
 import { TW } from "./generated/core_proto";
-import { WalletCore } from "./wallet-core";
+import { WalletCore } from "./src/wallet-core";
+import * as KeyStore from "./src/keystore";
 
 declare function load(): Promise<WalletCore>;
 
 export const initWasm: typeof load = Loader;
-export { TW, WalletCore };
-export { KeyStore } from "./keystore";
-export { ExtensionStorage } from "./extension-storage";
-export { FileSystemStorage } from "./fs-storage";
+export { TW, WalletCore, KeyStore };
