+++ wasm/src/keystore/index.ts
@@ -1,0 +1,10 @@
+// Copyright © 2017-2022 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+export { Default } from "./default-impl";
+export * from "./types";
+export { FileSystemStorage } from "./fs-storage";
+export { ExtensionStorage } from "./extension-storage";
